Unit Tests have been added

+++ Derivco.FibonacciSequence.Logic/Cache/FibonacciSequenceMemoryCacheService.cs
@@ -4,7 +4,7 @@ using Microsoft.Extensions.Configuration;
 
 namespace Derivco.FibonacciSequence.Logic.Cache
 {
-    public class FibonacciSequenceMemoryCacheService : MemoryCacheService<string, long>, IFibonacciSequenceMemoryCacheService
+    public class FibonacciSequenceMemoryCacheService : MemoryCacheService<string, uint>, IFibonacciSequenceMemoryCacheService
     {
         public FibonacciSequenceMemoryCacheService(IMemoryCache memoryCache, IConfiguration configuration) : base(memoryCache, configuration)
         {
