+++ Derivco.FibonacciSequence.Logic/Cache/IFibonacciSequenceMemoryCacheService.cs
@@ -2,7 +2,7 @@ using System.Numerics;
 
 namespace Derivco.FibonacciSequence.Logic.Cache
 {
-    public interface IFibonacciSequenceMemoryCacheService : IMemoryCacheService<string, long>
+    public interface IFibonacciSequenceMemoryCacheService : IMemoryCacheService<string, uint>
     {
         
     }
